Updated UI element cursor toggles & game manager
UI cursor gets toggled if any UI element is up that is interactable.

game manager is now the only object needed to be spawned

Added sound controls to the ui screens.

Toggling on and off the music will restart it, had to place music in a sound mixer for the volume control to work